refactor(api): rename import port and remove exportDate use in producer lastUpdate
